converter fixes

+++ main/converter.sh
@@ -13,10 +13,12 @@ if [ ! -d ~/converted_metadata_pickles ]; then
         cd -
 fi
 local_destination=~/raw_metadata_csvs
+S3_RAW_METADATA_DIR=s3://serratus-biosamples/bioprojects_csv/
+local_condensing_destination=~/converted_metadata_pickles
 echo "set up dirs"
 
-if [ "$1" == "-s" ]; then
-    S3_CONDENSED_METADATA_DIR=s3://serratus-biosamples/condense-test/
+if [[ "$1" == "-s" || "$1" == "-st" ]]; then
+    S3_CONDENSED_METADATA_DIR=s3://serratus-biosamples/condense-testing/
     if [ "$2" ]; then
         subset_size=$2
     else
@@ -24,45 +26,56 @@ if [ "$1" == "-s" ]; then
     fi
 
     # remove files if they exist
-    if [ -f file_list_with_sizes.txt ]; then
-        rm file_list_with_sizes.txt
-    fi
-    touch file_list_with_sizes.txt
-    size_list_file=file_list_with_sizes.txt
-    if [ -f cp_batch_command_list.txt ]; then
-        rm cp_batch_command_list.txt
-    fi
-    touch cp_batch_command_list.txt
-    cp_file_name=cp_batch_command_list.txt
+
     if [ -f csv_files_list.txt ]; then
         rm csv_files_list.txt
     fi
     touch csv_files_list.txt
     disk_files_name=csv_files_list.txt
 
-    s5cmd ls ${S3_RAW_METADATA_DIR} | awk '{print $NF, $(NF-1)}' | head -n ${subset_size} > ${size_list_file}
-    echo "completed getting file list via s5cmd ls"
+    if [ "$1" == "-s" ]; then
+        if [ -f file_list_with_sizes.txt ]; then
+            rm file_list_with_sizes.txt
+        fi
+        touch file_list_with_sizes.txt
+        size_list_file=file_list_with_sizes.txt
+        if [ -f cp_batch_command_list.txt ]; then
+            rm cp_batch_command_list.txt
+        fi
+        touch cp_batch_command_list.txt
+        cp_file_name=cp_batch_command_list.txt
 
-    while read -r line; do
-        file=$(echo ${line} | awk '{print $(NF-1)}')
-        echo "cp -f ${S3_RAW_METADATA_DIR}${file} ${local_destination}" >> ${cp_file_name}
-        echo "${local_destination}/${file}" >> ${disk_files_name}
-    done < ${size_list_file}
 
-    echo "completed creating cp batch command list"
+        s5cmd ls ${S3_RAW_METADATA_DIR} | sort -R | awk '{print $NF, $(NF-1)}' | head -n ${subset_size} > ${size_list_file}
+        echo "completed getting file list via s5cmd ls"
 
-    s5cmd run ${cp_file_name}
-    echo "completed copying raw csv files from s3 to local disk"
+        while read -r line; do
+            file=$(echo ${line} | awk '{print $(NF-1)}')
+            echo "cp -f ${S3_RAW_METADATA_DIR}${file} ${local_destination}" >> ${cp_file_name}
+            echo "${local_destination}/${file}" >> ${disk_files_name}
+        done < ${size_list_file}
 
-    time python3 converter_.py ${disk_files_name} ~/converted_metadata_pickles > log.txt
+        echo "completed creating cp batch command list"
+
+        s5cmd run ${cp_file_name}
+        echo "completed copying raw csv files from s3 to local disk"
+        rm ${cp_file_name}
+    else
+        # sample randomly directly from local disk
+        ls ${local_destination} | sort -R | head -n ${subset_size} > ${disk_files_name}
+        sed -i 's/^$local_destination\//' $disk_files_name
+        local_condensing_destination=~/converted_metadata_pickles_testing
+    fi
+
+
+    time python3 converter_.py ${disk_files_name} $local_condensing_destination > log.txt
     echo "completed converting csv files to pickle files of condensed metadata"
 
     # sync the converted_metadata_pickles with an s3 bucket called s3://serratus-biosamples/condensed-bioproject-metadata/
-    s5cmd sync ~/converted_metadata_pickles/ ${S3_CONDENSED_METADATA_DIR}
+    s5cmd sync $local_condensing_destination ${S3_CONDENSED_METADATA_DIR}
     echo "completed syncing pickle files to s3"
 
-    rm ${cp_file_name}
-    rm ${disk_files_name}
+    # rm ${disk_files_name}
 
 elif [ "$1" == "-f" ]; then
     S3_CONDENSED_METADATA_DIR=s3://serratus-biosamples/condensed-bioproject-metadata/
@@ -71,10 +84,10 @@ elif [ "$1" == "-f" ]; then
     echo "completed copying raw csv files from s3 to local disk"
 
     # time python3 converter_.py ${local_destination} ~/converted_metadata_pickles > log.txt
-    time parallel -j 4 python3 converter_.py ::: ${local_destination} ::: ~/converted_metadata_pickles > log.txt
+    time parallel -j 4 python3 converter_.py ::: ${local_destination} ::: $local_condensing_destination > log.txt
     echo "completed converting csv files to pickle files of condensed metadata"
 
-    s5cmd sync ~/converted_metadata_pickles/ ${S3_CONDENSED_METADATA_DIR}
+    s5cmd sync $local_local_condensing_destination ${S3_CONDENSED_METADATA_DIR}
     echo "completed syncing pickle files to s3"
 else
     echo "Please specify either -s for subset or -f for full"
